--- src/reflect/examples/simple_world/testing/__init__.py
@@ -1,11 +1,0 @@
-from reflect.examples.simple_world.testing.play_model import play_model
-from reflect.examples.simple_world.testing.play_real import play_real
-from reflect.examples.simple_world.testing.test_obs import test_obs
-from reflect.examples.simple_world.testing.play_agent import play_agent
-
-__all__ = [
-    "play_model",
-    "play_real",
-    "test_obs",
-    "play_agent"
-]
