--- src/reflect/examples/simple_world/training/__init__.py
@@ -1,11 +1,0 @@
-from reflect.examples.simple_world.training.train_obs import train_obs
-from reflect.examples.simple_world.training.train_wm import train_wm
-from reflect.examples.simple_world.training.train_stgrad import train_stgrad
-
-
-__all__ = [
-    "train_obs",
-    "train_wm",
-    "train_td3",
-    "train_stgrad"
-]
